Move the auto saving middleware into the Application slice.
